Fixed tests and backend imports after moving backend files and tests
